[KYUUBI #3893] [BUG] Fix spark extension: UnspecifiedDistribution does not have default partitioning.
### _Why are the changes needed?_

1. to fix #3893

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #3894 from FireFoxAhri/master.

Closes #3893

da15a000 [firefox] [KYUUBI #3893] [BUG] Fix spark extension: UnspecifiedDistribution does not have default partitioning.

Authored-by: firefox <309637962@qq.com>
(cherry picked from commit 8ef6494e4a887145b9cbb451ce7b63b435557fc1)
